refactor: improve code organization and type safety
- Extract status utilities and types into dedicated files
- Centralize WebSocket constants and improve type definitions
