+++ src/components/MessageCard.tsx
@@ -1,5 +1,6 @@
 import { Card, CardContent, CardHeader } from "./ui/card";
 import { Badge } from "./ui/badge";
+
 import type MetricMessage from "@/types/MetricMessage";
 
 interface MessageCardProps {
