+++ src/constants/ws.ts
@@ -1,0 +1,1 @@
+export const WEBHOOK_URL = "ws://localhost:3000";
