+++ src/types/ConnectionsStatus.ts
@@ -1,0 +1,5 @@
+export type ConnectionStatus =
+  | "connecting"
+  | "connected"
+  | "disconnected"
+  | "error";
